refactor : like, scrap을 post 패키지에서 분리
