+++ src/main/java/inu/codin/codin/domain/like/controller/LikeController.java
@@ -1,8 +1,8 @@
-package inu.codin.codin.domain.post.domain.like.controller;
+package inu.codin.codin.domain.like.controller;
 
 import inu.codin.codin.common.response.SingleResponse;
-import inu.codin.codin.domain.post.domain.like.dto.LikeRequestDto;
-import inu.codin.codin.domain.post.domain.like.service.LikeService;
+import inu.codin.codin.domain.like.dto.LikeRequestDto;
+import inu.codin.codin.domain.like.service.LikeService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
