+++ src/main/java/inu/codin/codin/domain/like/dto/LikeRequestDto.java
@@ -1,6 +1,6 @@
-package inu.codin.codin.domain.post.domain.like.dto;
+package inu.codin.codin.domain.like.dto;
 
-import inu.codin.codin.domain.post.domain.like.entity.LikeType;
+import inu.codin.codin.domain.like.entity.LikeType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -10,7 +10,7 @@ import lombok.Getter;
 public class LikeRequestDto {
 
     @NotNull
-    @Schema(description = "좋아요를 반영할 entity 타입(POST, COMMENT, REPLY)", example = "POST")
+    @Schema(description = "좋아요를 반영할 entity 타입(POST, COMMENT, REPLY, REVIEW)", example = "POST")
     private LikeType likeType;
 
     @NotBlank
