+++ src/main/java/inu/codin/codin/domain/like/exception/LikeCreateFailException.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.post.domain.like.exception;
+package inu.codin.codin.domain.like.exception;
 
 public class LikeCreateFailException extends RuntimeException{
     public LikeCreateFailException(String message) {
