+++ src/main/java/inu/codin/codin/domain/like/exception/LikeRemoveFailException.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.post.domain.like.exception;
+package inu.codin.codin.domain.like.exception;
 
 public class LikeRemoveFailException extends RuntimeException {
     public LikeRemoveFailException(String message) {
