+++ src/main/java/inu/codin/codin/domain/scrap/controller/ScrapController.java
@@ -1,7 +1,7 @@
-package inu.codin.codin.domain.post.domain.scrap.controller;
+package inu.codin.codin.domain.scrap.controller;
 
 import inu.codin.codin.common.response.SingleResponse;
-import inu.codin.codin.domain.post.domain.scrap.service.ScrapService;
+import inu.codin.codin.domain.scrap.service.ScrapService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
