+++ src/main/java/inu/codin/codin/domain/scrap/exception/ScrapCreateFailException.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.post.domain.scrap.exception;
+package inu.codin.codin.domain.scrap.exception;
 
 public class ScrapCreateFailException extends RuntimeException {
     public ScrapCreateFailException(String message) {
