+++ src/main/java/inu/codin/codin/domain/user/service/UserService.java
@@ -5,11 +5,11 @@ import inu.codin.codin.common.security.util.SecurityUtils;
 import inu.codin.codin.domain.email.repository.EmailAuthRepository;
 import inu.codin.codin.domain.post.domain.comment.entity.CommentEntity;
 import inu.codin.codin.domain.post.domain.comment.repository.CommentRepository;
-import inu.codin.codin.domain.post.domain.like.entity.LikeEntity;
-import inu.codin.codin.domain.post.domain.like.entity.LikeType;
-import inu.codin.codin.domain.post.domain.like.repository.LikeRepository;
-import inu.codin.codin.domain.post.domain.scrap.entity.ScrapEntity;
-import inu.codin.codin.domain.post.domain.scrap.repository.ScrapRepository;
+import inu.codin.codin.domain.like.entity.LikeEntity;
+import inu.codin.codin.domain.like.entity.LikeType;
+import inu.codin.codin.domain.like.repository.LikeRepository;
+import inu.codin.codin.domain.scrap.entity.ScrapEntity;
+import inu.codin.codin.domain.scrap.repository.ScrapRepository;
 import inu.codin.codin.domain.post.dto.response.PostPageResponse;
 import inu.codin.codin.domain.post.entity.PostEntity;
 import inu.codin.codin.domain.post.repository.PostRepository;
